Conflicts:
	scilab/Version.incl
	scilab/modules/action_binding/sci_gateway/cpp/sci_notify.cpp
	scilab/modules/api_scilab/src/cpp/api_boolean.cpp
	scilab/modules/api_scilab/src/cpp/api_boolean_sparse.cpp
	scilab/modules/api_scilab/src/cpp/api_common.cpp
	scilab/modules/api_scilab/src/cpp/api_double.cpp
	scilab/modules/api_scilab/src/cpp/api_handle.cpp
	scilab/modules/api_scilab/src/cpp/api_int.cpp
	scilab/modules/api_scilab/src/cpp/api_list.cpp
	scilab/modules/api_scilab/src/cpp/api_pointer.cpp
	scilab/modules/api_scilab/src/cpp/api_poly.cpp
	scilab/modules/api_scilab/src/cpp/api_sparse.cpp
	scilab/modules/api_scilab/src/cpp/api_string.cpp
	scilab/modules/core/src/c/callinterf.c
	scilab/modules/core/src/c/run.c
	scilab/modules/core/src/c/stack2.c
	scilab/modules/differential_equations/sci_gateway/fortran/bbvode.f
	scilab/modules/differential_equations/sci_gateway/fortran/bfeval.f
	scilab/modules/differential_equations/sci_gateway/fortran/bjac.f
	scilab/modules/differential_equations/sci_gateway/fortran/bjacd.f
	scilab/modules/differential_equations/sci_gateway/fortran/bresd.f
	scilab/modules/differential_equations/sci_gateway/fortran/bsurf.f
	scilab/modules/differential_equations/sci_gateway/fortran/bsurfd.f
	scilab/modules/differential_equations/sci_gateway/fortran/bydot.f
	scilab/modules/differential_equations/sci_gateway/fortran/bydot2.f
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_eye.f
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_ones.f
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_rand.f
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_zeros.f
	scilab/modules/external_objects_java/sci_gateway/c/gw_external_objects_java.c
	scilab/modules/fileio/tests/unit_tests/mgeti.dia.ref
	scilab/modules/fileio/tests/unit_tests/mgeti.tst
	scilab/modules/fileio/tests/unit_tests/mputi.dia.ref
	scilab/modules/fileio/tests/unit_tests/mputi.tst
	scilab/modules/graphic_objects/Makefile.am
	scilab/modules/graphic_objects/Makefile.in
	scilab/modules/graphics/Makefile.in
	scilab/modules/graphics/graphics.vcxproj
	scilab/modules/graphics/graphics.vcxproj.filters
	scilab/modules/graphics/includes/gw_graphics.h
	scilab/modules/graphics/sci_gateway/c/gw_graphics.c
	scilab/modules/graphics/src/c/getHandleProperty/GetHashTable.c
	scilab/modules/graphics/src/c/getHandleProperty/getHandleProperty.h
	scilab/modules/graphics/src/c/getHandleProperty/get_segs_color_property.c
	scilab/modules/graphics/src/c/getHandleProperty/get_type_property.c
	scilab/modules/gui/Makefile.am
	scilab/modules/gui/Makefile.in
	scilab/modules/gui/includes/gw_gui.h
	scilab/modules/gui/sci_gateway/c/gw_gui.c
	scilab/modules/gui/src/c/gui.vcxproj
	scilab/modules/gui/src/c/gui.vcxproj.filters
	scilab/modules/helptools/data/pages/homepage-en_US.html
	scilab/modules/helptools/data/pages/homepage-fr_FR.html
	scilab/modules/integer/src/fortran/i_mput.f
	scilab/modules/localization/sci_gateway/c/sci_addlocalizationdomain.c
	scilab/modules/scicos_blocks/src/c/bouncexy.c
	scilab/modules/scicos_blocks/src/c/canimxy.c
	scilab/modules/scicos_blocks/src/c/canimxy3d.c
	scilab/modules/scicos_blocks/src/c/cevscpe.c
	scilab/modules/scicos_blocks/src/c/cfscope.c
	scilab/modules/scicos_blocks/src/c/cmat3d.c
	scilab/modules/scicos_blocks/src/c/cmatview.c
	scilab/modules/scicos_blocks/src/c/cmscope.c
	scilab/modules/scicos_blocks/src/c/cscope.c
	scilab/modules/scicos_blocks/src/c/cscopxy.c
	scilab/modules/scicos_blocks/src/c/cscopxy3d.c
	scilab/modules/scicos_blocks/src/fortran/sciblk.f
	scilab/modules/signal_processing/sci_gateway/fortran/bgetx.f
	scilab/modules/signal_processing/sci_gateway/fortran/bgety.f
	scilab/modules/spreadsheet/sci_gateway/c/sci_csvDefault.c
	scilab/modules/spreadsheet/sci_gateway/c/sci_csvRead.c
	scilab/modules/spreadsheet/src/c/checkCsvWriteFormat.c
	scilab/modules/spreadsheet/src/c/csvRead.c
	scilab/modules/statistics/tests/nonreg_tests/bug_3640.dia.ref
	scilab/modules/statistics/tests/nonreg_tests/bug_3640.tst
	scilab/modules/string/sci_gateway/c/sci_length.c
	scilab/modules/xcos/includes/gw_xcos.h
	scilab/modules/xcos/sci_gateway/c/gw_xcos.c
	scilab/modules/xml/sci_gateway/cpp/sci_xmlXPath.cpp
	scilab/modules/xml/src/cpp/XMLList.cpp
	scilab/modules/xml/src/cpp/XMLObject.cpp
	scilab/modules/xml/src/cpp/XMLObject.hxx
	scilab/modules/xml/src/cpp/XMLRhsValue.hxx

+++ scilab/modules/fileio/src/c/mgeti.h
@@ -35,7 +35,7 @@ int SWAP(char type[], int *fd);
  * @param char type[]
  * @param ierr
  */
-FILEIO_IMPEXP void C2F(mgeti) (int *fd, int *res, int *n, char type[], int *ierr);
+FILEIO_IMPEXP void C2F(mgeti) (int *fd, long long *res, int *n, char type[], int *ierr);
 
 #endif /* __MGETI_H__ */
